requirements.txt
pdfplumber
pytesseract
pillow
rich
